feat: add some features
1. add uid field
2. sleep when encouter errors

+++ app/models/user_dynamic.py
@@ -13,7 +13,7 @@ class UserDynamic(Base):
     type_id = Column(Integer)
     oid = Column(BIGINT)
     # 用户uid
-    uid = Column(BIGINT)
+    uid = Column(Integer)
     # 爬取状态
     status = Column(Integer)
     ctime = Column(Integer)
